tele_lab/requirements.txt
Django==2.0.4
django-webpack-loader==0.6.0
djangorestframework==3.9.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
django-oauth-toolkit==1.2.0
django-braces==1.13.0